+++ tests/testthat/test-lintr.R
@@ -1,15 +1,10 @@
-linters_config <- lintr::with_defaults(
-  object_usage_linter = NULL,
-  line_length_linter = lintr::line_length_linter(120),
-  camel_case_linter = NULL,
-  snake_case_linter = NULL,
-  multiple_dots_linter = NULL
-)
+context("lintr")
 
-if (requireNamespace("lintr", quietly = TRUE)) {
-  context("lints")
-  test_that("Package Style", {
-
-    lintr::expect_lint_free(linters = linters_config)
-  })
-}
+test_that("lintr", {
+  print(getwd())
+  lintr::expect_lint_free(
+    path = "../../00_pkg_src/semantic.dashboard",
+    relative_path = TRUE,
+    lintr::line_length_linter(150)
+  )
+})